added primitive logout functionality
