+++ src/app/login/page.tsx
@@ -1,22 +1,35 @@
 "use client";
 
-import { FormEvent, useState } from "react";
-import Link from "next/link";
+import { useState } from "react";
+import { isEmail, useForm } from "@mantine/form";
+import { TextInput, Group, Button, PasswordInput } from "@mantine/core";
 import { useRouter } from "next/navigation";
 import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
 import { app } from "../../firebase";
 
+interface LoginFormValues {
+  email: string;
+  password: string;
+}
+
 export default function Login() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
   const [error, setError] = useState("");
   const router = useRouter();
 
-  async function handleSubmit(event: FormEvent) {
-    event.preventDefault();
-    setError("");
+  const loginForm = useForm({
+    initialValues: {
+      email: "",
+      password: "",
+    },
+    validate: {
+      email: isEmail("Invalid email"),
+    },
+  });
 
+  async function handleSubmit(values: LoginFormValues) {
+    setError("");
     try {
+      const { email, password } = values;
       const credential = await signInWithEmailAndPassword(
         getAuth(app),
         email,
@@ -37,79 +50,29 @@ export default function Login() {
   }
 
   return (
-    <main className="flex min-h-screen flex-col items-center justify-center p-8">
-      <div className="w-full bg-white rounded-lg shadow dark:border md:mt-0 sm:max-w-md xl:p-0 dark:bg-gray-800 dark:border-gray-700">
-        <div className="p-6 space-y-4 md:space-y-6 sm:p-8">
-          <h1 className="text-xl font-bold leading-tight tracking-tight text-gray-900 md:text-2xl dark:text-white">
-            Speak thy secret word!
-          </h1>
-          <form
-            onSubmit={handleSubmit}
-            className="space-y-4 md:space-y-6"
-            action="#"
-          >
-            <div>
-              <label
-                htmlFor="email"
-                className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
-              >
-                Your email
-              </label>
-              <input
-                type="email"
-                name="email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                id="email"
-                className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                placeholder="name@company.com"
-                required
-              />
-            </div>
-            <div>
-              <label
-                htmlFor="password"
-                className="block mb-2 text-sm font-medium text-gray-900 dark:text-white"
-              >
-                Password
-              </label>
-              <input
-                type="password"
-                name="password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                id="password"
-                placeholder="••••••••"
-                className="bg-gray-50 border border-gray-300 text-gray-900 sm:text-sm rounded-lg focus:ring-primary-600 focus:border-primary-600 block w-full p-2.5 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                required
-              />
-            </div>
-            {error && (
-              <div
-                className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative"
-                role="alert"
-              >
-                <span className="block sm:inline">{error}</span>
-              </div>
-            )}
-            <button
-              type="submit"
-              className="w-full text-white bg-gray-600 hover:bg-gray-700 focus:ring-4 focus:outline-none focus:ring-primary-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-gray-600 dark:hover:bg-gray-700 dark:focus:ring-primary-800"
-            >
-              Enter
-            </button>
-            <p className="text-sm font-light text-gray-500 dark:text-gray-400">
-              Don&apos;t have an account?{" "}
-              <Link
-                href="/register"
-                className="font-medium text-gray-600 hover:underline dark:text-gray-500"
-              >
-                Register here
-              </Link>
-            </p>
-          </form>
+    <form onSubmit={loginForm.onSubmit((values) => handleSubmit(values))}>
+      <TextInput
+        withAsterisk
+        label="Email"
+        placeholder="knguyen@gmail.com"
+        {...loginForm.getInputProps("email")}
+      />
+      <PasswordInput
+        label="Password"
+        placeholder="password"
+        {...loginForm.getInputProps("password")}
+      />
+      <Group mt="md">
+        <Button type="submit">Submit</Button>
+      </Group>
+      {error && (
+        <div
+          className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded relative"
+          role="alert"
+        >
+          <span className="block sm:inline">{error}</span>
         </div>
-      </div>
-    </main>
+      )}
+    </form>
   );
 }
