+++ src/pages/api/index.ts
@@ -1,0 +1,1 @@
+// TODO: Can we move all the login, register and, logout endpoints here like in the starter example: https://github.com/awinogrodzki/next-firebase-auth-edge/blob/main/examples/next-typescript-starter/api/index.ts
